inference/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
opencv-python>=4.8.0
mediapipe>=0.10.0
pillow>=10.0.0
numpy>=1.24.0
captum>=0.6.0
matplotlib>=3.7.0
aiofiles>=23.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0